pilates-booking/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==5.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0
flake8==6.1.0
python-dotenv==1.0.0
celery==5.3.4
stripe==8.7.0
factory_boy==3.3.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==20.1.0
aiosqlite==0.19.0